Rename QuestionsContainer to src/Quiz
